+++ task/Results.js
@@ -26,7 +26,7 @@ var Results;
         //    throw new Error("cannot create a result with both a non-null value and non-null error");
         //}
         return {
-            hasResult: value != null,
+            hasResult: error == null,
             result: value,
             error: error
         };