+++ test/TaskTest.ts
@@ -1,7 +1,7 @@
 ﻿import chai = require("chai");
 import mocha = require("mocha");
 import Task = require("../task/Task");
-import TaskStatus = require("../task/TaskStatus");
+import TaskState = require("../task/TaskState");
 
 
 var asr = chai.assert;
@@ -11,15 +11,15 @@ suite("Task", function TaskTest() {
 
     test("task-success", function taskSuccessTest(done) {
         // test success
-        var res1 = new Task("a", () => "a-success");
+        var res1 = Task.newTask("a", () => "a-success");
 
         asr.equal(res1.name, "a");
-        asr.equal(res1.status, TaskStatus.CREATED);
+        asr.equal(res1.state, TaskState.CREATED);
 
         res1.getPromise().then((res) => {
             asr.equal(res, "a-success");
             asr.equal(res1.getResult(), "a-success");
-            asr.equal(res1.status, TaskStatus.COMPLETED);
+            asr.equal(res1.state, TaskState.COMPLETED);
             done();
         }, (err) => {
             asr.equal(true, false, "task " + res1.name + " failed unexpectedly");
@@ -32,10 +32,10 @@ suite("Task", function TaskTest() {
 
     test("task-failure", function taskFailureTest(done) {
         // test failure
-        var res1 = new Task("a", () => { throw "a-error"; });
+        var res1 = Task.newTask("a", () => { throw "a-error"; });
 
         asr.equal(res1.name, "a");
-        asr.equal(res1.status, TaskStatus.CREATED);
+        asr.equal(res1.state, TaskState.CREATED);
 
         res1.getPromise().then((res) => {
             asr.equal(true, false, "task " + res1.name + " succeeded unexpectedly");
@@ -43,7 +43,7 @@ suite("Task", function TaskTest() {
         }, (err) => {
             asr.equal(err, "a-error");
             asr.equal(res1.getError(), "a-error");
-            asr.equal(res1.status, TaskStatus.ERRORED);
+            asr.equal(res1.state, TaskState.ERRORED);
             done();
         });